main_project/requirements.txt
absl-py==0.2.2
appdirs==1.4.0
argon2-cffi==18.3.0
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
cachetools==2.1.0
cffi==1.11.5
colorama==0.3.7
coverage==4.3.4
cycler==0.10.0
decorator==4.3.0
Django==2.1.2
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
entrypoints==0.2.3
gast==0.2.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
grpcio==1.12.1
html5lib==0.9999999
httplib2==0.11.3
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
mistune==0.8.3
mysql-connector-python==8.0.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.4
oauth2client==4.1.3
packaging==16.8
pandas==0.23.1
pandocfilters==1.4.2
parso==0.2.1
pbs==0.110
pickleshare==0.7.4
Pillow==5.3.0
prompt-toolkit==1.0.15
protobuf==3.6.0
py==1.4.32
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyflakes==1.5.0
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==2.6.0
pytz==2018.4
pywin32==220
pywinpty==0.5.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.12.5
rsa==4.0
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sklearn==0.0
style==1.1.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
update==0.0.1
uritemplate==3.0.0
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
